# pkg/docs/swagger.yaml
basePath: /api
definitions:
  controllers.httpError:
    properties:
      error:
        type: string
    type: object
  controllers.httpSuccess:
    properties:
      result:
        type: string
    type: object
  model.Album:
    properties:
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    required:
    - title
    - userId
    type: object
  model.Comment:
    properties:
      body:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      postId:
        type: integer
    required:
    - body
    - email
    - name
    - postId
    type: object
  model.Photo:
    properties:
      albumId:
        type: integer
      id:
        type: integer
      thumbnailUrl:
        type: string
      title:
        type: string
      url:
        type: string
    required:
    - albumId
    - thumbnailUrl
    - title
    - url
    type: object
  model.Post:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    required:
    - body
    - title
    - userId
    type: object
  model.Todo:
    properties:
      completed:
        type: boolean
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    required:
    - completed
    - title
    - userId
    type: object
  model.User:
    properties:
      address:
        properties:
          city:
            type: string
          geo:
            properties:
              lat:
                type: string
              lng:
                type: string
            required:
            - lat
            - lng
            type: object
          street:
            type: string
          suite:
            type: string
          zipcode:
            type: string
        required:
        - city
        - geo
        - street
        - suite
        - zipcode
        type: object
      company:
        properties:
          bs:
            type: string
          catchPhrase:
            type: string
          name:
            type: string
        required:
        - bs
        - catchPhrase
        - name
        type: object
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      username:
        type: string
      website:
        type: string
    required:
    - address
    - company
    - email
    - name
    - phone
    - username
    - website
    type: object
...
host: cao-veio:5000
info:
  contact:
    name: Fabio Azevedo
  description: API REST created to apply knowledge in learning the Go language
  license:
    name: MIT
    url: https://www.mit.edu/~amini/LICENSE.md
  termsOfService: http://swagger.io/terms/
  title: CRUD API REST
  version: 1.0.0
paths:
...
  /v1/albums/{id}:
    delete:
      consumes:
      - application/json
      description: delete album by id
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/controllers.httpSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Delete Album By ID
      tags:
      - albums
    get:
      consumes:
      - application/json
      description: get album by id
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Album'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Get Album By ID
      tags:
      - albums
...
  /v1/comments/{id}:
    delete:
      consumes:
      - application/json
      description: delete comment by id
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/controllers.httpSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Delete Comment By ID
      tags:
      - comments
    get:
      consumes:
      - application/json
      description: get comment by id
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Comment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Get Comment By ID
      tags:
      - comments
...
  /v1/photos/{id}:
    delete:
      consumes:
      - application/json
      description: delete photo by id
      parameters:
      - description: Photo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/controllers.httpSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Delete Photo By ID
      tags:
      - photos
    get:
      consumes:
      - application/json
      description: get photo by id
      parameters:
      - description: Photo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Photo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Get Photo By ID
      tags:
      - photos
...
  /v1/posts/{id}:
    delete:
      consumes:
      - application/json
      description: delete post by id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/controllers.httpSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Delete Post By ID
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: get post by id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Get Post By ID
      tags:
      - posts
...
  /v1/todos/{id}:
    delete:
      consumes:
      - application/json
      description: delete todo by id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/controllers.httpSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Delete Todo By ID
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: get todo by id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Get Todo By ID
      tags:
      - todos
...
  /v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/controllers.httpSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Delete User By ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: get user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.httpError'
      summary: Get User By ID
      tags:
      - users
swagger: "2.0"